Move resize from lib.rs to individual printers
