+++ src/printer/kitty.rs
@@ -1,9 +1,10 @@
 use crate::error::{ViuError, ViuResult};
 use crate::printer::Printer;
+use crate::utils::{fit_dimensions, terminal_size};
 use console::{Key, Term};
 use crossterm::cursor::{MoveRight, MoveTo, MoveToPreviousLine};
 use crossterm::execute;
-use image::GenericImageView;
+use image::{DynamicImage, GenericImageView};
 use lazy_static::lazy_static;
 use std::io::Write;
 
@@ -144,16 +145,14 @@ fn print_local(img: &image::DynamicImage, config: &crate::Config) -> ViuResult {
     }
 
     // get the desired width and height
-    // TODO: create a calc_dimensions function that does the same without actually modifying an image.
-    //  Also, the h returend is 2* what we want, because of blocks' height
-    let (w, h) = crate::resize(img, config.width, config.height).dimensions();
+    let (w, h) = resize(&img, config.width, config.height);
 
     print!(
         "\x1b_Gf=32,s={},v={},c={},r={},a=T,t=t;{}\x1b\\",
         img.width(),
         img.height(),
         w,
-        h / 2,
+        h,
         base64::encode(path.to_str().unwrap())
     );
     println!();
@@ -162,6 +161,51 @@ fn print_local(img: &image::DynamicImage, config: &crate::Config) -> ViuResult {
     Ok(())
 }
 //TODO default_features false of console
+
+// Figure out the best dimensions for the printed image, based on user's input.
+// Returns the dimensions of how the image should be printed in terminal cells.
+fn resize(img: &DynamicImage, width: Option<u32>, height: Option<u32>) -> (u16, u16) {
+    let (img_width, img_height) = img.dimensions();
+
+    // The height of a terminal cell is twice its width. Thus, when calling fit_dimensions,
+    //  it has to know that the actual dimensions are (w, 2*h), so that it can correctly
+    //  maintain the aspect ratio.
+    //  The return height, however, represents terminal cells and has to be divided by two.
+    //
+    // For example, to fit image 160x80 in a terminal grid 80x24:
+    // Aspect ratio is 160/80 = 2:1,
+    // the terminal can be split into 80x46 equivalent squares, and
+    // the image has to be scaped down. Preserving aspect ratio,
+    // the maximum fit will be 80x40 squares, or 80x20 terminal cells.
+    // Also, after 1 extra row is kept (see below), the returned dimensions will be 80x19.
+
+    // Match user's width and height preferences
+    match (width, height) {
+        (None, None) => {
+            let (term_w, term_h) = terminal_size();
+            let (w, h) = fit_dimensions(img_width, img_height, term_w as u32, 2 * term_h as u32);
+
+            //TODO: is -1 row always necessary? Maybe only when term_h*2 == h?
+            // One less row because two reasons:
+            // - the prompt after executing the command will take a line
+            // - gifs flicker
+            (w as u16, (h / 2 - 1) as u16)
+        }
+        // Either width or height is specified, will fit and preserve aspect ratio.
+        (Some(w), None) => {
+            let (w, h) = fit_dimensions(img_width, img_height, w, 2 * img_height);
+            (w as u16, (h / 2) as u16)
+        }
+        (None, Some(h)) => {
+            let (w, h) = fit_dimensions(img_width, img_height, img_width, 2 * h);
+            (w as u16, (h / 2) as u16)
+        }
+        (Some(w), Some(h)) => {
+            // Both width and height are specified, will resize to match exactly
+            (w as u16, h as u16)
+        }
+    }
+}
 
 // Create a file in temporary dir and write the byte slice to it.
 // Since the file is persisted, the user is responsible for deleting it afterwards.
@@ -175,4 +219,92 @@ fn store_in_tmp_file(raw_img: &[u8]) -> std::result::Result<std::path::PathBuf, 
     tmpfile.write_all(raw_img).unwrap();
     tmpfile.flush().unwrap();
     Ok(path)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn get_large_test_image() -> DynamicImage {
+        DynamicImage::ImageRgba8(image::RgbaImage::new(600, 500))
+    }
+
+    fn get_small_test_image() -> DynamicImage {
+        DynamicImage::ImageRgba8(image::RgbaImage::new(40, 25))
+    }
+
+    #[test]
+    fn test_resize_none() {
+        let width = None;
+        let height = None;
+
+        let img = get_large_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 57);
+        assert_eq!(h, 23);
+
+        let img = get_small_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 76);
+        assert_eq!(h, 23);
+
+        let img = DynamicImage::ImageRgba8(image::RgbaImage::new(160, 80));
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 80);
+        assert_eq!(h, 19);
+    }
+
+    #[test]
+    fn test_resize_some_none() {
+        let width = Some(100);
+        let height = None;
+
+        let img = get_large_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 100);
+        assert_eq!(h, 41);
+
+        let img = get_small_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 80);
+        assert_eq!(h, 25);
+
+        let width = Some(6);
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 6);
+        assert_eq!(h, 1);
+    }
+
+    #[test]
+    fn test_resize_none_some() {
+        let width = None;
+        let height = Some(90);
+
+        let img = get_large_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 216);
+        assert_eq!(h, 90);
+
+        let height = Some(4);
+        let img = get_small_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 12);
+        assert_eq!(h, 4);
+    }
+
+    #[test]
+    fn test_resize_some_some() {
+        let width = Some(15);
+        let height = Some(9);
+
+        let img = get_large_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 15);
+        assert_eq!(h, 9);
+
+        let img = get_small_test_image();
+        let (w, h) = resize(&img, width, height);
+        assert_eq!(w, 15);
+        assert_eq!(h, 9);
+    }
 }
